+++ tests/ecc.rs
@@ -40,10 +40,7 @@ fn component_add_point() {
     }
 
     impl Circuit for TestCircuit {
-        fn circuit<C>(&self, composer: &mut C) -> Result<(), Error>
-        where
-            C: Composer,
-        {
+        fn circuit(&self, composer: &mut Composer) -> Result<(), Error> {
             let w_p1 = composer.append_point(self.p1);
             let w_p2 = composer.append_point(self.p2);
             let w_sum = composer.append_point(self.sum);
@@ -144,7 +141,7 @@ fn component_mul_generator() {
     }
 
     impl Circuit for TestCircuit {
-        fn circuit<C: Composer>(&self, composer: &mut C) -> Result<(), Error> {
+        fn circuit(&self, composer: &mut Composer) -> Result<(), Error> {
             let w_scalar = composer.append_witness(self.scalar);
             let w_result = composer.append_point(self.result);
 
@@ -246,10 +243,7 @@ fn component_mul_point() {
     }
 
     impl Circuit for TestCircuit {
-        fn circuit<C>(&self, composer: &mut C) -> Result<(), Error>
-        where
-            C: Composer,
-        {
+        fn circuit(&self, composer: &mut Composer) -> Result<(), Error> {
             let w_scalar = composer.append_witness(self.scalar);
             let w_point = composer.append_point(self.point);
             let w_result = composer.append_point(self.result);